+++ CHANGELOG.md
@@ -5,7 +5,7 @@
 - Android Components/Drop-in
   version: [5.14.0](https://docs.adyen.com/online-payments/release-notes/#releaseNote=2025-08-19-android-componentsdrop-in-5.14.0).
 - iOS Components/Drop-in
-  version: [5.19.2](https://docs.adyen.com/online-payments/release-notes/?title%5B0%5D=iOS+Components%2FDrop-in#releaseNote=2025-07-14-ios-componentsdrop-in-5.19.2).
+  version: [5.20.0](https://docs.adyen.com/online-payments/release-notes/?title%5B0%5D=iOS+Components%2FDrop-in#releaseNote=2025-07-14-ios-componentsdrop-in-5.19.2).
 
 ### Improved